Refactored code to include all models into a seperate project. All projects now include the models as a dependency.
