--- services/api-gateway/src/main/java/api_gateway/services/images/models/Image.java
@@ -1,17 +1,0 @@
-package api_gateway.services.images.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Image {
-
-	private String image;
-
-    public String getImage() {
-		return image;
-	}
-
-	public void setImage(String name) {
-		this.image = name;
-	}
-}
