+++ services/api-gateway/src/main/java/api_gateway/services/movie/MovieIntegrationService.java
@@ -7,7 +7,7 @@ import org.springframework.web.client.RestTemplate;
 import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import com.netflix.hystrix.contrib.javanica.command.ObservableResult;
 
-import api_gateway.services.movie.models.Movie;
+import edu.cmu.ini.ericsson.practicum.models.movieService.Movie;
 import rx.Observable;
 
 
@@ -29,8 +29,8 @@ public class MovieIntegrationService {
 
     private Movie stubMovie(final String mID) {
         Movie stub = new Movie();
-        stub.setDataReleased("01-Jan-2001");
-        stub.setImdbURI("www.imdb.com");
+        stub.setReleaseDate("01-Jan-2001");
+        stub.setImdbUrl("www.imdb.com");
         stub.setName("Default Movie");
         return stub;
     }
