+++ services/eureka-service/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: ${SERVICE_PORT}
+  port: ${SERVICE_PORT:8761}
 
 eureka:
   instance:
