--- services/frontend-service/src/main/java/frontend_service/models/Rating.java
@@ -1,35 +1,0 @@
-package frontend_service.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Rating {
-
-    private String userId;
-    private String movieId;
-    private int rating;
-    
-	public String getUserId() {
-		return userId;
-	}
-	
-	public void setUserId(String userId) {
-		this.userId = userId;
-	}
-	
-	public String getMovieId() {
-		return movieId;
-	}
-	
-	public void setMovieId(String movieId) {
-		this.movieId = movieId;
-	}
-	
-	public int getRating() {
-		return rating;
-	}
-	
-	public void setRating(int rating) {
-		this.rating = rating;
-	}
-}
