+++ services/movie-service/src/main/java/movie_service/domain/MovieRepository.java
@@ -3,6 +3,8 @@ package movie_service.domain;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
+import edu.cmu.ini.ericsson.practicum.models.movieService.Movie;
+
 @RepositoryRestResource(collectionResourceRel = "movie", path = "movie")
 public interface MovieRepository extends MongoRepository<Movie, String> {
